PWN-7376 - changes according to comments EthTokenConverter.kt added ethMetadataToToken removed from TokenConverter.kt
