+++ ethereumkit/src/main/java/org/p2p/ethereumkit/external/repository/EthereumRepository.kt
@@ -1,10 +1,10 @@
 package org.p2p.ethereumkit.external.repository
 
 import java.math.BigInteger
-import org.p2p.ethereumkit.external.model.EthTokenMetadata
+import org.p2p.core.token.Token
 
 interface EthereumRepository {
     fun init(seedPhrase: List<String>)
     suspend fun getBalance(): BigInteger
-    suspend fun loadWalletTokens(): List<EthTokenMetadata>
+    suspend fun loadWalletTokens(): List<Token.Eth>
 }
